Added resolvers and getUserId

+++ src/resolvers/Mutation.js
@@ -1,32 +1,76 @@
+import bcrypt from 'bcryptjs'
+import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
+
 const Mutation = {
-    async createUser(parent, args, { prisma },info) {
-        const emailExists = await prism.exists.User({ email: email: args.data.email })
+    async createUser(parent, args, { prisma }, info) {
         // data validation is actually not needed since Prisma does the job for us
-        if(emailExists) {
-            throw new Error("User already exists!")
+        // const emailExists = await prisma.exists.User({ email: args.data.email })
+        // if(emailExists) {
+        //     throw new Error("User already exists!")
+        // }
+
+        // take in password -> validate password -> hash password -> genereate auth token
+        if(args.data.password.length < 8) {
+            throw new Error("Pasword must be 8 characters or longer")
         }
-        return prisma.mutation.createUser({ data: args.data }, info)
+
+        const password = await bcrypt.hash(args.data.password, 10)
+
+        const user = prisma.mutation.createUser({ 
+            data: {
+                ...args.data,
+                password
+            } 
+        })
+        return { 
+            user, 
+            token: jwt.sign({ userId: user.id }, 'thisisasecret')
+        }
     },
-    async deleteUser(paret, args, { prisma }, info) {
-        const userExists = await prisma.exists.User({ id: args.id })
-        if(!userExists) {
-            throw new Err("User not found")
+    async login(parent, args, { prisma }, info) {
+        const user = await prisma.query.user({
+            where: {
+                email: args.data.email
+            }
+        })
+        if(!user) {
+            throw new Error("Unable to login")
         }
+        const isMatch = await bcrypt.compare(args.data.password, user.password)
+        if(!isMatch) {
+            throw new Error("Unable to login")
+        }
+        return {
+            user,
+            token: jwt.sign({ userId: user.id }, 'thisisasecret')
+        }
+
+    },
+    async deleteUser(paret, args, { prisma, request }, info) {
+        // no need to check, Prisma handles that verification with the @unique directive 
+        // const userExists = await prisma.exists.User({ id: args.id })
+        // if(!userExists) {
+        //     throw new Err("User not found")
+        // }
+        const userId = getUserId(request)
         return prisma.mutation.deleteUser({ 
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     }, 
-    updateUser(parent, args, { prisma }, info){
+    updateUser(parent, args, { prisma, request }, info){
+        const userId = getUserId(request)
         return prisma.mutation.updateUser({ 
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
     },
-    createPost(parent, args, { prisma }, info){
+    createPost(parent, args, { prisma, request }, info){
+        const userId = getUserId(request)
         return prisma.mutation.createPost({
             data: {
                 title: args.data.title,
@@ -34,16 +78,22 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }
         }, info)
     },
-    async deletePost(parent, args, { prisma }, info){
-        const postExists = await prisma.exists.Post({ id: args.id })
+    async deletePost(parent, args, { prisma, request }, info){
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }   
+        })
         if(!postExists) {
-            throw new Error("Post not found!")
+            throw new Error("Unable to delete post!")
         }
         return prisma.mutation.deletePost({
             where: {
@@ -51,21 +101,55 @@ const Mutation = {
             }
         }, info)
     },
-    updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if(!postExists) {
+            throw new Error("Unable to update the post")
+        }
+
+        const isPublished = prisma.exists.Post({
+            id: args.id,
+            published: true
+        })
+
+        if(isPublished && args.data.published === false) {
+            await prisma.mutation.deleteManyComments({
+                where: {
+                    post: {
+                        id: args.id
+                    }
+                }
+            })
+        }
         return prisma.mutation.updatePost({
             where: {
                 id: args.id
             },
-            data: data
+            data: args.data
         }, info)
     },
-    createComment(parent, args, { prisma }, info){
+    async createComment(parent, args, { prisma, request }, info){
+        const userId = getUserId(request)
+        const postExists = prisma.exists.Post({
+            id: args.data.post,
+            published: true
+        })
+
+        if(!postExists) {
+            throw new Error("Unable to find post")
+        }
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -76,15 +160,35 @@ const Mutation = {
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if(!commentExists) {
+            throw new Error("Comment could not be updated")
+        }
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
             },
-            data: data
+            data: args.data
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info){
+    deleteComment(parent, args, { prisma, request }, info){
+        const userId = getUserId(request)
+        const commentExists = prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if(!commentExists) {
+            throw new Error("Comment could not be updated")
+        } 
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
@@ -93,3 +197,5 @@ const Mutation = {
     }
 
 }
+
+export default Mutation
